Changed the way the scripture text is added.
This new methods will allow unlimited references, and it will be more optimized. I'm thinking that it would be better to remove the database all the fields relating to the scripture text, and truncate all those fields into one field labeled just "scripture". All the scripture references can then be stored in that one field in the format: "Genesis 1:1-5


git-svn-id: http://joomlacode.org/svn/biblestudy/BibleStudy/Trunk@190 a73faa7a-dd35-0410-af13-bd1c84eb1a9d