Update biblestudy.podcast.class.php
I added a routine to check for -1 in the array (text) returned from the podcast_id. We had set the default in the db to be -1 and no way to get that out of the field once saved (another bug). It ends up returning all episodes from every podcast that has -1 in it.