Update sermons.php
The default_order parameter was not being used. Discovered that the $params variable in the list query was not being set to the template params. I changed the $params to getstate(mergedparams) then changed what is being looked for - no longer a number but the actual text ASC or DESC. Now the default order works, and users can override that with the drop down on the list menu.